rename clp-s to clp_s, add benchall
