+++ assets/clp_s/Dockerfile
