+++ assets/clp_s/main.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 import sys
-from template import DATASETS_PATH, WORK_DIR, Benchmark, logger
+from src.template import DATASETS_PATH, WORK_DIR, Benchmark, logger
 
 CLP_OUT_PATH = f"{WORK_DIR}/archives"
 CLP_S_BINARY = "/clp/clp-s"
